Add a member 'registerp' to ecs-entity.
If registerp is t, it indicates that the entity is registered in the
global entity list *entity-list*. This addition is because the function
find-the-entity is frequently used and heavy (profiled in JavaScript
version). So this member is added for like a cache of the result.

Note: The previous find-the-entity used find-a-entity. Probably,
improving the performance of the find-a-entity itself will be required.