Add frame-promise for asynchronous processing
The frame-promise is inspired from JavaScript's Promise. It is checked
that each frame-promise is resolved in first of frame (ecs-main). If
it has been resoloved, following frame-promises are invoked.

In addition, func-with-pred that is base of frame-promise is updated
to predicate (and execute) in order of registration because I think it
is more natural that chain of "then" is invoked in order of
registration.