Fix the issue about directly referring package internal variable
In delete-ecs-entity, it replaces *entity-list* by a new list.
So if we simply refer cl-ps-ecs.ecs::*entity-list* (in the JavaScript,
it becomes clPsEcs_ecs._internal.ENTITYLIST), such change cannot be
reflected to the referrer in JavaScript side.

To avoid it, in the do-ecs-entities macro, we use a new function to
refer *entity-list* rather than use directly *entity-list*.

A similar issue can be occured in do-ecs-systems, we applied a similar
fix to it.